# perf: Optimize hash joins with an empty build side