## Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

N/A.

## Rationale for this change

<!--
 Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
 Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.  
-->

When executing hash joins, the build side is first built from the left relation and then the right relation is joined with it. However, when the build side has no rows, the join operation can be mostly skipped, improving performance.

For example, here is a simple anti join query, where `t1` has 100M rows and `t2` has none:
```sql
SELECT *
FROM t1
LEFT ANTI JOIN t2 on t1.k = t2.k
```

Here is the hash join operator in the current implementation:
```txt
HashJoinExec: mode=Partitioned, join_type=RightAnti, on=[(k@0, k@0)], metrics=[
    output_rows=100000000,
    build_input_batches=0,
    build_input_rows=0,
    input_batches=11733,
    input_rows=100000000,
    output_batches=23403,
    build_mem_used=876,
    build_time=2.8693ms,
    join_time=216.251396302s
]
```

And here is the optimized hash join operation:
```txt
HashJoinExec: mode=Partitioned, join_type=RightAnti, on=[(k@0, k@0)], metrics=[
    output_rows=100000000,
    build_input_batches=0,
    build_input_rows=0,
    input_batches=11733,
    input_rows=100000000,
    output_batches=11733,
    build_mem_used=876,
    build_time=2.4597ms,
    join_time=36.038306ms
]
```

The total join time went from 216s to just 36ms.


## What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

- Changed `process_probe_batch` in `physical-plan/hash_join.rs` to optimized the join.
- Added multiple sqllogictests.

## Are these changes tested?

<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example, are they covered by existing tests)?
-->

Yes.

## Are there any user-facing changes?

<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

No.

<!--
If there are any breaking changes to public APIs, please add the `api change` label.
-->