@jonathanc-n 2025-07-09
> @jonathanc-n Since after #16434 the hash map is not directly accessible, I've added an `is_empty` method to `JoinHashMapType`. Please check if this is the preferred approach.

Yes this looks good