@jonathanc-n 2025-07-09
@nuno-faria We can return early from collect_left_input after intaking batches and checking the number of batches 
```rust

if batches.len() == 0 {
        return Ok(JoinLeftData::new(
            Box::new(JoinHashMapU32::with_capacity(0)),
            RecordBatch::new_empty(schema),
            Vec::new(),
            Mutex::new(BooleanBufferBuilder::new(0)),
            AtomicUsize::new(probe_threads_count),
            reservation,
        ));
    };

```